FBCModelTests.jl functions

FBCModelTests.FBCModelTests — Module

module FBCModelTests

A collection of tests for constraint-based flux metabolic models.

FROG

FBCModelTests.FROG — Module

module FROG

A collection of reproducibility checks for constraint-based metabolic models together with a report generator and tester.

See https://www.ebi.ac.uk/biomodels/curation/fbc for details.

FBCModelTests.FROG.ObjectiveValue — Type

FBCModelTests.FROG.FROGMetadata — Type

mutable struct Dict{String, Any} <: AbstractDict{String, Any}

FBCModelTests.FROG.FROGObjectiveReport — Type

struct FROGObjectiveReport

Fields

optimum::Union{Nothing, Float64}

reactions::Dict{String, FBCModelTests.FROG.FROGReactionReport}

gene_deletions::Dict{String, Union{Nothing, Float64}}

FBCModelTests.FROG.FROGReactionReport — Type

struct FROGReactionReport

Fields

objective_flux::Union{Nothing, Float64}

fraction_optimum::Float64

variability_min::Union{Nothing, Float64}

variability_max::Union{Nothing, Float64}

deletion::Union{Nothing, Float64}

FBCModelTests.FROG.FROGReportData — Type

mutable struct Dict{String, FBCModelTests.FROG.FROGObjectiveReport} <: AbstractDict{String, FBCModelTests.FROG.FROGObjectiveReport}

FBCModelTests.FROG.compare_reports — Method

compare_reports(
    report_dir_a::String,
    report_dir_b::String;
    kwargs...
) -> Union{Test.FallbackTestSet, Test.DefaultTestSet}

A simple wrapper for comparing 2 previously generated reports in their respective directories. Additional arguments are fowarded to ReportTests.test_report_compatibility.

FBCModelTests.FROG.generate_report — Method

generate_report(
    filename::String;
    report_dir,
    optimizer,
    modifications,
    workers,
    basefilename
)

A complete function for one-shot generation of FROG reports. Use ReportGenerators.generate_report_data, ReportGenerators.generate_metadata and ReportIO.save_report for finer control of the process.

Reading and writing reports

FBCModelTests.FROG.ReportIO — Module

module ReportIO

Functions for reading and writing FROG reports.

FBCModelTests.FROG.ReportIO.load_report — Method

load_report(
    report_dir::String
) -> NamedTuple{(:metadata, :report), Tuple{Dict{String, Any}, Dict{String, FBCModelTests.FROG.FROGObjectiveReport}}}

Reverse of save_report.

FBCModelTests.FROG.ReportIO.save_report — Method

save_report(
    r::Dict{String, FBCModelTests.FROG.FROGObjectiveReport},
    metadata::Dict{String, Any};
    report_dir,
    basefilename
)

Write the contents of FROGReportData to the 4 TSV files as specified by FROG standard, and additionally write the metadata into the JSON file.

Generating and testing the reports

FBCModelTests.FROG.ReportGenerators — Module

module ReportGenerators

Functions for generating FROG report data and metadata file contents.

FBCModelTests.FROG.ReportGenerators.ResetObjective — Type

struct ResetObjective <: COBREXA.ModelWrapper

Fields

model::COBREXA.MetabolicModel

objective::SparseArrays.SparseVector{Float64, Int64}

FBCModelTests.FROG.ReportGenerators.frog_objective_report — Function

frog_objective_report(sbml_model::COBREXA.SBMLModel)
frog_objective_report(
    sbml_model::COBREXA.SBMLModel,
    objective::Union{Nothing, String};
    optimizer,
    modifications,
    workers,
    fraction_optimum
)

Generate a FROGObjectiveReport containing the reproducibility data for a single objective in the SBML model.

FBCModelTests.FROG.ReportGenerators.generate_metadata — Method

generate_metadata(filename::String; optimizer, basefilename)

FBCModelTests.FROG.ReportGenerators.generate_report_data — Method

generate_report_data(
    model::COBREXA.SBMLModel;
    kwargs...
) -> Dict{Any, Any}

Generate FROGReportData for a model.

FBCModelTests.FROG.ReportTests — Module

module ReportTests

Function for testing the compatibility of FROG report data.

FBCModelTests.FROG.ReportTests.test_metadata_compatibility — Method

test_metadata_compatibility(
    a::Dict{String, Any},
    b::Dict{String, Any}
)

FBCModelTests.FROG.ReportTests.test_report_compatibility — Method

test_report_compatibility(
    a::Dict{String, FBCModelTests.FROG.FROGObjectiveReport},
    b::Dict{String, FBCModelTests.FROG.FROGObjectiveReport};
    absolute_tolerance,
    relative_tolerance
) -> Union{Test.FallbackTestSet, Test.DefaultTestSet}

MEMOTE

FBCModelTests.Memote — Module

module Memote

This package contains a collection of tests based on Memote. See Lieven, C., Beber, M.E., Olivier, B.G. et al. MEMOTE for standardized genome-scale metabolic model testing. Nat Biotechnol 38, 272–276 (2020). https://doi.org/10.1038/s41587-020-0446-y for details.

FBCModelTests.Memote.generate_report — Method

generate_report(
    model::COBREXA.MetabolicModel,
    optimizer;
    config
) -> Dict

Generate a machine-readable report of model properties that can be used as a baseline for checking the quality of the model. Suitable for being exported as JSON.

FBCModelTests.Memote.generate_report — Method

generate_report(
    filename::String,
    optimizer;
    kwargs...
) -> Dict

Overload of generate_report that works directly on a given filename.

FBCModelTests.Memote.run_tests — Method

run_tests(
    model::COBREXA.MetabolicModel,
    optimizer;
    config,
    filename
) -> Union{Test.FallbackTestSet, Test.DefaultTestSet}

Run a MEMOTE-like test suite on model using optimizer to solve optimization problems; some basic parameters and thresholds are taken from config.

Some of the tests internally convert the input model to StandardModel of COBREXA; supplying a StandardModel may thus increase efficiency of the whole process.

FBCModelTests.Memote.run_tests — Method

run_tests(
    filename::String,
    optimizer;
    kwargs...
) -> Union{Test.FallbackTestSet, Test.DefaultTestSet}

Overload of run_tests that works directly with a file.

MEMOTE utilities

FBCModelTests.Memote.Utils — Module

module  Utils

Miscellaneous functions used by memote style tests, not typically user facing.

FBCModelTests.Memote.Utils._compare_flux_bounds — Method

_compare_flux_bounds(
    fluxes,
    bound,
    tol,
    comparison_operator
) -> Dict{String, Tuple{String, Float64}}

Internal helper function to compare fluxes with specific bounds.

FBCModelTests.Memote.Utils._has_sensible_gpr — Method

_has_sensible_gpr(model::COBREXA.MetabolicModel, rid) -> Any

Internal helper function that determines if a reaction has a gene reaction rule and that each gene in the rule is contained in the model.

FBCModelTests.Memote.Utils._probably_transport_reaction — Method

_probably_transport_reaction(
    model::COBREXA.MetabolicModel,
    rid,
    test_annotation
) -> Bool

Determine if a reaction is probably a transport reaction by checking if:

it has sbo annotations corresponding to a transport reaction

the reaction contains metabolites from at least 2 different compartments

if at least 1 metabolite does not undergo a chemical transformation (via formula or annotation checks)

Note, PTS type transport reactions will be missed if they do not have sbo annotations. This test may yield false negatives.

FBCModelTests.Memote.Utils.median_bounds — Method

median_bounds(
    model::COBREXA.MetabolicModel
) -> Tuple{Any, Any}

A helper function that returns the median upper and lower bounds in a tuple. If none can be calculated, constants from COBREXA are used as the default values.

FBCModelTests.Memote.Utils.to_element — Method

to_element(x::String) -> PeriodicTable.Element

Return the chemical element of x.

FBCModelTests.Memote.Config — Module

module Config

Module housing the configuration parameters for the memote-style tests.

FBCModelTests.Memote.Config.AnnotationConfig — Type

mutable struct AnnotationConfig

Fields

gene_annotation_keywords::Vector{String}

gene_annotation_regexes::Dict{String, Regex}

metabolite_annotation_keywords::Vector{String}

metabolite_annotation_regexes::Dict{String, Regex}

reaction_annotation_keywords::Vector{String}

reaction_annotation_regexes::Dict{String, Regex}

minimum_fraction_database_conformity::Float64

minimum_fraction_database_annotations::Float64

FBCModelTests.Memote.Config.BasicConfig — Type

mutable struct BasicConfig

Fields

minimum_metabolic_coverage::Float64

minimum_growth_rate::Float64

maximum_growth_rate::Float64

optimizer_modifications::Vector{Function}

FBCModelTests.Memote.Config.BiomassConfig — Type

mutable struct BiomassConfig

Fields

biomass_strings::Vector{String}

growth_metabolites::Dict{String, String}

ignored_precursors::Vector{String}

essential_precursors::Dict{String, String}

minimum_growth_rate::Float64

optimizer_modifications::Vector{Function}

FBCModelTests.Memote.Config.ConsistencyConfig — Type

mutable struct ConsistencyConfig

Fields

mass_ignored_reactions::Vector{String}

charge_ignored_reactions::Vector{String}

consistency_ignored_reactions::Vector{String}

tolerance_threshold::Float64

FBCModelTests.Memote.Config.EnergyConfig — Type

mutable struct EnergyConfig

Fields

atpm_strings::Vector{String}

energy_dissipating_metabolites::Dict{String, String}

additional_energy_generating_reactions::Vector{COBREXA.Reaction}

ignored_energy_reactions::Vector{String}

optimizer_modifications::Vector{Function}

FBCModelTests.Memote.Config.GPRAssociationConfig — Type

mutable struct GPRAssociationConfig

Fields

test_annotation::String

FBCModelTests.Memote.Config.MemoteConfig — Type

mutable struct MemoteConfig

A grouping of parameters used by the metabolic testing infrastructure.

Fields

annotation::FBCModelTests.Memote.Config.AnnotationConfig

basic::FBCModelTests.Memote.Config.BasicConfig

biomass::FBCModelTests.Memote.Config.BiomassConfig

consistency::FBCModelTests.Memote.Config.ConsistencyConfig

energy::FBCModelTests.Memote.Config.EnergyConfig

gpra::FBCModelTests.Memote.Config.GPRAssociationConfig

metabolite::FBCModelTests.Memote.Config.MetaboliteConfig

network::FBCModelTests.Memote.Config.NetworkConfig

reaction::FBCModelTests.Memote.Config.ReactionConfig

FBCModelTests.Memote.Config.MetaboliteConfig — Type

mutable struct MetaboliteConfig

Fields

formula_corner_cases::Vector{String}

charge_corner_cases::Vector{Int64}

medium_only_imported::Bool

test_annotation::String

FBCModelTests.Memote.Config.NetworkConfig — Type

mutable struct NetworkConfig

Fields

condition_number::Float64

fva_bound::Float64

cycle_tol::Float64

minimum_metabolite_flux::Float64

optimizer_modifications::Vector{Function}

FBCModelTests.Memote.Config.ReactionConfig — Type

mutable struct ReactionConfig

Fields

test_annotation::String

ignore_annotations::Vector{String}

bound_default::Float64

MEMOTE checks

Basic checks

FBCModelTests.Memote.Basic — Module

module Basic

Basic metabolic model tests.

FBCModelTests.Memote.Basic.model_compartments — Method

model_compartments(model::COBREXA.MetabolicModel) -> Set

Return the number of unique compartments in the model.

FBCModelTests.Memote.Basic.model_has_compartments — Method

model_has_compartments(
    model::COBREXA.MetabolicModel
) -> Bool

Test if the model has one or more compartments.

FBCModelTests.Memote.Basic.model_has_genes — Method

model_has_genes(model::COBREXA.MetabolicModel) -> Any

Test if the model has any genes.

FBCModelTests.Memote.Basic.model_has_metabolites — Method

model_has_metabolites(model::COBREXA.MetabolicModel) -> Any

Test if the model has any metabolites.

FBCModelTests.Memote.Basic.model_has_name — Method

model_has_name(model::COBREXA.MetabolicModel) -> Bool

Test if the model has an id. Note, this test will fail unless you use a StandardModel and its id field is assigned.

FBCModelTests.Memote.Basic.model_has_reactions — Method

model_has_reactions(model::COBREXA.MetabolicModel) -> Any

Test if the model has any reactions.

FBCModelTests.Memote.Basic.model_metabolic_coverage — Method

model_metabolic_coverage(
    model::COBREXA.MetabolicModel
) -> Any

Calculate the metabolic coverage by dividing the number of reactions by the number of genes.

FBCModelTests.Memote.Basic.model_metabolic_coverage_exceeds_minimum — Method

model_metabolic_coverage_exceeds_minimum(
    model::COBREXA.MetabolicModel;
    config
) -> Any

Test if the metabolic coverage, calculated with model_metabolic_coverage, exceeds config.basic.minimum_metabolic_coverage.

FBCModelTests.Memote.Basic.model_solves_in_default_medium — Method

model_solves_in_default_medium(
    model::COBREXA.MetabolicModel,
    optimizer;
    config
) -> Any

Check if the model can be solved under default conditions and yield a reasonable growth rate. Here reasonable is set via config.basic.minimum_growth_rate and config.basic.maximum_growth_rate. Optionally, pass optimization modifications to the solver through config.basic.optimizer_modifications.

Consistency checks

FBCModelTests.Memote.Consistency — Module

module Consistency

This module checks if the metabolic model is overall consistent.

FBCModelTests.Memote.Consistency.find_disconnected_metabolites — Method

find_disconnected_metabolites(
    model::COBREXA.MetabolicModel
) -> Any

Returns a list of all metabolites that aren't part of any reactions.

FBCModelTests.Memote.Consistency.model_is_consistent — Method

model_is_consistent(
    model::COBREXA.MetabolicModel,
    optimizer;
    config
) -> Bool

Determines if the model is stoichiometrically consistent. Note, stoichiometric consistency does not guarantee that mass balances must hold in the model. A more robust check is reactions_mass_unbalanced, but this works if not all metabolites have mass assigned to them. Based on Gevorgyan, Albert, Mark G. Poolman, and David A. Fell. "Detection of stoichiometric inconsistencies in biomolecular models." Bioinformatics (2008). Optionally ignore some reactions in this analysis by adding reaction IDs to config.consistency.consistency_ignored_reactions.

FBCModelTests.Memote.Consistency.reactions_charge_unbalanced — Method

reactions_charge_unbalanced(
    model::COBREXA.MetabolicModel;
    config
) -> Vector{String}

Iterates through all the reactions in model and checks if the charges across each reaction balance. Returns a list of reaction IDs that are charge unbalanced. Optionally, use config.consistency.mass_ignored_reactions to pass a vector of reaction ids to ignore in this process. Internally biomass and exchang reactions are ignored.

FBCModelTests.Memote.Consistency.reactions_mass_unbalanced — Method

reactions_mass_unbalanced(
    model::COBREXA.MetabolicModel;
    config
) -> Vector{String}

Iterates through all the reactions in model and checks if the mass across each reaction balances. Returns a list of reaction IDs that are mass unbalanced. Optionally, use config.consistency.charge_ignored_reactions to pass a vector of reaction ids to ignore in this process. Internally biomass and exchang reactions are ignored.

FBCModelTests.Memote.Consistency.unbounded_flux_in_default_medium — Function

unbounded_flux_in_default_medium(
    model::COBREXA.MetabolicModel,
    fva_result::Tuple{Any, Any}
) -> Tuple{Dict{String, Tuple{String, Float64}}, Dict{String, Tuple{String, Float64}}}
unbounded_flux_in_default_medium(
    model::COBREXA.MetabolicModel,
    fva_result::Tuple{Any, Any},
    config
) -> Tuple{Dict{String, Tuple{String, Float64}}, Dict{String, Tuple{String, Float64}}}

Finds reactions that can carry unlimited flux under default conditions. The function compares the fluxes of the model calculated by FVA (using flux_variability_analysis_dict from COBREXA) with the median bounds of the model (+/- treshold) and returns the fluxes which are grearter or smaller than the bounds in a two seperate dictionaries.

Metabolite checks

FBCModelTests.Memote.Metabolite — Module

module Metabolite

A module testing various metabolite properties.

FBCModelTests.Memote.Metabolite.metabolites_are_duplicated — Method

metabolites_are_duplicated(
    model::COBREXA.MetabolicModel,
    m1,
    m2;
    config
) -> Any

Test if metabolites m1 and m2 are different by comparing their config.metabolite.test_annotation field in the annotations of each metabolite. Note, if no annotations are present for one or both of the metabolites, then return true.

FBCModelTests.Memote.Metabolite.metabolites_duplicated_in_compartment — Method

metabolites_duplicated_in_compartment(
    model::COBREXA.MetabolicModel;
    config
) -> Dict{String, Set{String}}

Return a dictionary of metabolites that are duplicated in their compartment.

FBCModelTests.Memote.Metabolite.metabolites_medium_components — Method

metabolites_medium_components(
    model::COBREXA.MetabolicModel;
    config
) -> Vector{String}

Return a list of all boundary reactions that allow flux into the model and create a metabolite. Assume that boundary reactions only create a single metabolite. Use the testing config, config.metabolite.only_imported = false, to also return metabolites that can be produced by the model under default conditions.

FBCModelTests.Memote.Metabolite.metabolites_no_charge — Method

metabolites_no_charge(
    model::COBREXA.MetabolicModel;
    config
) -> Any

List all metabolites without a charge. Use config.metabolite.charge_corner_cases to specify an extra case to check for charge's that are not properly assigned.

FBCModelTests.Memote.Metabolite.metabolites_no_formula — Method

metabolites_no_formula(
    model::COBREXA.MetabolicModel;
    config
) -> Any

List all metabolites without a formula. Use config.metabolite.formula_corner_cases to specify an extra case to check for formula's that are not properly assigned.

FBCModelTests.Memote.Metabolite.metabolites_unique — Method

metabolites_unique(
    model::COBREXA.MetabolicModel;
    config
) -> Set{String}

Return a list of unique metabolites in model. Uses metabolites_are_duplicated internally and forwards test_annotation to it. The latter argument is used to determine if two metabolites are the same by checking for any correspondence.

Reaction checks

FBCModelTests.Memote.Reaction — Module

module Reaction

A module testing reaction properties.

FBCModelTests.Memote.Reaction._is_constrained — Method

_is_constrained(lb, ub, default) -> Union{Missing, Bool}

Test if a reaction is constrained.

FBCModelTests.Memote.Reaction.duplicate_reactions — Method

duplicate_reactions(
    model::COBREXA.MetabolicModel
) -> Set{String}

Return a list of all reactions that are duplicated.

FBCModelTests.Memote.Reaction.find_all_purely_metabolic_reactions — Method

find_all_purely_metabolic_reactions(
    model::COBREXA.MetabolicModel;
    config
) -> Tuple{Set{String}, Set{String}}

Find all purely metabolic reactions and return two sets, one with unconstrained metabolic reactions (directional constraints do not count as constraints in this sense), and one with constrained metabolic reactions. For a constraint to be not purely directional, the lower or upper bound needs to be different from [-config.reactions.bound_default, 0, config.reactions.bound_default]. Metabolic reactions exclude transport, boundary and biomass reactions.

FBCModelTests.Memote.Reaction.find_all_transport_reactions — Method

find_all_transport_reactions(
    model::COBREXA.MetabolicModel;
    config
) -> Tuple{Set{String}, Set{String}}

Find all transport reactions and and return two sets, one with unconstrained transport reactions (directional constraints do not count as constraints in this sense), and one with constrained transport reactions. For a constraint to be not purely directional, the lower or upper bound needs to be different from [-config.reactions.bound_default, 0, config.reactions.bound_default]. Transport reactions are heuristically identified, see Memote.Utils._probably_transport_reaction, which partially uses reaction annotations. Set the annotation field to use via config.reactions.rest_annotation.

FBCModelTests.Memote.Reaction.reactions_with_identical_genes — Method

reactions_with_identical_genes(
    model::COBREXA.MetabolicModel
) -> Dict{Vector{String}, Set{String}}

Identify reactions that have the same gene reaction rules. Does not take directionality or compartments into account.

FBCModelTests.Memote.Reaction.reactions_with_partially_identical_annotations — Method

reactions_with_partially_identical_annotations(
    model::COBREXA.MetabolicModel;
    config
) -> Set{String}

Find all reactions with overlapping annotation information. Internally calls COBREXA.annotation_index. Some annotations, like sbo terms will necessarily be non-unique, ignore annotations like this by editing config.reaction.ignore_annotations.

GPRAssociation checks

FBCModelTests.Memote.GPRAssociation — Module

module GPRAssociation

A module testing various facets of the gene reaction associations.

FBCModelTests.Memote.GPRAssociation.reactions_transport_no_gpr — Method

reactions_transport_no_gpr(
    model::COBREXA.MetabolicModel;
    config
) -> Any

Return a list of transport reactions that do not have gene reaction rules assigned to them.

FBCModelTests.Memote.GPRAssociation.reactions_with_complexes — Method

reactions_with_complexes(
    model::COBREXA.MetabolicModel
) -> Any

Return a list of reaction ids that have protein complexes assigned to them.

FBCModelTests.Memote.GPRAssociation.reactions_without_gpr — Method

reactions_without_gpr(model::COBREXA.MetabolicModel) -> Any

Return a list of reaction ids that do not have gene reaction rules (aka gene protein reaction associations).

Biomass checks

FBCModelTests.Memote.Biomass — Module

module Biomass

This module contains tests that check the consistency of the biomass reaction.

FBCModelTests.Memote.Biomass.atp_present_in_biomass — Method

atp_present_in_biomass(
    model::COBREXA.MetabolicModel;
    config
) -> Any

Check if the biomass reaction consumes ATP and H₂O, and produces ADP, HO₄P, and H⁺. Each of these metabolites have a lookup table mapping them to the name space of the model, defined in config.biomass.growth_metabolites. These need to be set if you use anything other than the BiGG namespace.

FBCModelTests.Memote.Biomass.biomass_missing_essential_precursors — Method

biomass_missing_essential_precursors(
    model::COBREXA.MetabolicModel;
    config
) -> Dict{String, Vector{String}}

Tests if each biomass reaction contains a set of essential precursors, listed in config.biomass.essential_precursors. Note, this function only works on a lumped biomass function.

FBCModelTests.Memote.Biomass.find_blocked_biomass_precursors — Method

find_blocked_biomass_precursors(
    model::COBREXA.MetabolicModel,
    optimizer;
    config
) -> Dict{String, Vector{String}}

Check if the model can synthesize all of the biomass precursors in all of the biomass functions, except those listed in config.biomass.ignored_precursors in the default medium. Set any optimizer modifications with config.biomass.optimizer_modifications.

FBCModelTests.Memote.Biomass.model_biomass_is_consistent — Method

model_biomass_is_consistent(
    model::COBREXA.MetabolicModel;
    config
) -> Any

Check that the molar mass of each biomass reactions falls within [1 - 1e-3, 1 + 1e-6] by calling model_biomass_molar_mass internally.

FBCModelTests.Memote.Biomass.model_biomass_molar_mass — Method

model_biomass_molar_mass(
    model::COBREXA.MetabolicModel;
    config
) -> Any

For each biomass reaction, identified by model_biomass_reactions, calculate the molar weight of the reaction by summing the products of the associated metabolite coefficients with their molar masses.

FBCModelTests.Memote.Biomass.model_biomass_reactions — Method

model_biomass_reactions(
    model::COBREXA.MetabolicModel;
    config
) -> Set

Identify all the biomass reactions in the model using both sbo annotations, as well biomass strings typically contained in their reaction IDs. Use config.biomass.biomass_strings to update the list of strings to look for.

Network checks

FBCModelTests.Memote.Network — Module

module Network

A module testing the network and topology properties of the model.

FBCModelTests.Memote.Network._find_orphan_or_deadend_metabolites — Method

_find_orphan_or_deadend_metabolites(
    model::COBREXA.MetabolicModel;
    consumed
) -> Vector{String}

Helper function to find orphan or deadend metabolites. Specify consumed=true to consider orphan metabolites or false to consider deadend metabolites. Set complete_medium=true to open all boundary reactions to simulate a complete medium.

FBCModelTests.Memote.Network.find_all_universally_blocked_reactions — Method

find_all_universally_blocked_reactions(
    model::COBREXA.MetabolicModel,
    optimizer;
    config
) -> Any

Make all boundary reactions reversible and run FVA on the model to find all reactions that are universally blocked. Optimizer modifications can be passed through config.network.optimizer_modifications

FBCModelTests.Memote.Network.find_complete_medium_orphans_and_deadends — Method

find_complete_medium_orphans_and_deadends(
    model::COBREXA.MetabolicModel,
    optimizer;
    config
) -> Dict{Symbol, Vector{String}}

Return a dictionary mapping orphan and deadend metabolites that occur in the model in complete medium. Complete medium is modeled by opening all the bounday reactions. For each metabolite FBA is run with a temporary reaction either consuming or producing the metabolite in question. At minimum a flux of config.network.minimum_metabolite_flux must be attained for the metabolite to pass the test.

FBCModelTests.Memote.Network.find_cycle_reactions — Method

find_cycle_reactions(
    model::COBREXA.MetabolicModel,
    optimizer;
    config
) -> Set{String}

Find all reactions that participate in stoichiometrically balanced cycles by closing all boundary reactions and running fva on the resultant model.

FBCModelTests.Memote.Network.find_deadend_metabolites — Method

find_deadend_metabolites(
    model::COBREXA.MetabolicModel
) -> Vector{String}

Find all metabolites that can only (excludes reversible reactions) be produced in the model by inspecting the stoichiometric matrix.

FBCModelTests.Memote.Network.find_orphan_metabolites — Method

find_orphan_metabolites(
    model::COBREXA.MetabolicModel
) -> Vector{String}

Find all metabolites that can only (excludes reversible reactions) be consumed in the model by inspecting the stoichiometric matrix.

FBCModelTests.Memote.Network.stoichiometric_matrix_is_well_conditioned — Method

stoichiometric_matrix_is_well_conditioned(
    model::COBREXA.MetabolicModel;
    config
) -> Any

Test if the stoichiometric matrix is well conditioned by determining if stoichiometric_max_min_ratio is less than 10⁹ (which can be set in config.network.condition_number).

FBCModelTests.Memote.Network.stoichiometric_max_min_ratio — Method

stoichiometric_max_min_ratio(
    model::COBREXA.MetabolicModel
) -> Any

Return the ratio of the absolute maximum and minimum value of the nonzero coefficients in the stoichiometric matrix of model.

Annotation checks

FBCModelTests.Memote.Annotation — Module

module Annotation

This module contains tests used to check the coverage and conformance of reaction, gene, and metabolite annotations.

FBCModelTests.Memote.Annotation._find_nonconformal_components — Method

_find_nonconformal_components(
    model::COBREXA.MetabolicModel,
    id_accessor,
    annotation_accessor,
    annotation_regex
) -> Dict{String, Vector{String}}

Helper function to find all the annotations that do not conform in the model.

FBCModelTests.Memote.Annotation._find_unannotated_components — Method

_find_unannotated_components(
    model::COBREXA.MetabolicModel,
    id_accessor,
    annotation_accessor,
    annotation_kws
) -> Dict{String, Vector{String}}

Helper function to find all the unannotated components in the model.

FBCModelTests.Memote.Annotation.find_all_unannotated_genes — Method

find_all_unannotated_genes(
    model::COBREXA.MetabolicModel
) -> Any

Checks if every gene has an annotation and returns an array of genes that do not have annotatons.

FBCModelTests.Memote.Annotation.find_all_unannotated_metabolites — Method

find_all_unannotated_metabolites(
    model::COBREXA.MetabolicModel
) -> Any

Checks if every metabolite has an annotation and returns an array of metabolites that do not have annotatons.

FBCModelTests.Memote.Annotation.find_all_unannotated_reactions — Method

find_all_unannotated_reactions(
    model::COBREXA.MetabolicModel
) -> Any

Checks if every reaction has an annotation and returns an array of reactions that do not have annotatons.

FBCModelTests.Memote.Annotation.find_database_unannotated_genes — Method

find_database_unannotated_genes(
    model;
    config
) -> Dict{String, Vector{String}}

Checks if the databases listed in config.annotation.gene_annotation_keywords are present in the gene annotations. Returns a dictionary of annotation keywords mapped to a list of genes that do not include the keyword.

FBCModelTests.Memote.Annotation.find_database_unannotated_metabolites — Method

find_database_unannotated_metabolites(
    model;
    config
) -> Dict{String, Vector{String}}

Checks if the databases listed in config.annotation.metabolite_annotation_keywords are present in the metabolite annotations. Returns a dictionary of annotation keywords mapped to a list of metabolites that do not include the keyword.

FBCModelTests.Memote.Annotation.find_database_unannotated_reactions — Method

find_database_unannotated_reactions(
    model;
    config
) -> Dict{String, Vector{String}}

Checks if the databases listed in config.annotation.reaction_annotation_keywords are present in the reaction annotations. Returns a dictionary of annotation keywords mapped to a list of reactions that do not include the keyword.

FBCModelTests.Memote.Annotation.find_nonconformal_gene_annotations — Method

find_nonconformal_gene_annotations(
    model;
    config
) -> Dict{String, Vector{String}}

Check if the gene annotation entry conforms to commonly recognized formats of annotation database using regex patterns. Uses the database formats listed in config.annotation.gene_annotation_regexes to test the conformity. Returns a dictionary mapping the database id to a list of genes that do not conform.

FBCModelTests.Memote.Annotation.find_nonconformal_metabolite_annotations — Method

find_nonconformal_metabolite_annotations(
    model;
    config
) -> Dict{String, Vector{String}}

Check if the metabolite annotation entry conforms to commonly recognized formats of annotation database using regex patterns. Uses the database formats listed in config.annotation.metabolite_annotation_regexes to test the conformity. Returns a dictionary mapping the database id to a list of metabolites that do not conform.

FBCModelTests.Memote.Annotation.find_nonconformal_reaction_annotations — Method

find_nonconformal_reaction_annotations(
    model;
    config
) -> Dict{String, Vector{String}}

Check if the reaction annotation entry conforms to commonly recognized formats of annotation database using regex patterns. Uses the database formats listed in config.annotation.reaction_annotation_regexes to test the conformity. Returns a dictionary mapping the database id to a list of reactions that do not conform.

Energy checks

FBCModelTests.Memote.Energy — Module

module Energy

This module checks if the model is energetically sensible.

FBCModelTests.Memote.Energy.model_has_atpm_reaction — Method

model_has_atpm_reaction(
    model::COBREXA.MetabolicModel;
    config
) -> Union{Missing, Bool}

Check if model has an ATP maintenance reaction built in (also called a non-growth associated maintenance cost). Looks for reaction annotations corresponding to the sbo maintenance term, or looks for reaction ids that contain the strings listed in config.energy.atpm_strings.

FBCModelTests.Memote.Energy.model_has_no_erroneous_energy_generating_cycles — Method

model_has_no_erroneous_energy_generating_cycles(
    model::COBREXA.MetabolicModel,
    optimizer;
    config
) -> Bool

Attempts to detect if the model contains any erroneous energy generating cycles by closing all exchange reactions and using flux balance analysis to maximize the sum of fluxes through a set of energy dissipating reactions. The flux sum should be zero if the model is free of energy generating reactions. This function is based on Fritzemeier, Claus Jonathan, et al. "Erroneous energy-generating cycles in published genome scale metabolic networks: Identification and removal." PLoS computational biology (2017). The energy dissipating reactions are based on the source paper, and include:

ATP + H2O --> ADP + H + Phosphate
CTP + H2O --> CDP + H + Phosphate
GTP + H2O --> GDP + H + Phosphate
UTP + H2O --> UDP + H + Phosphate
ITP + H2O --> IDP + H + Phosphate
NADH --> H + NAD
NADPH --> H + NADP
FADH2 --> 2 H + FAD
FMNH2 --> 2 H + FMN
Ubiquinol-8 --> 2 H + Ubiquinone-8
Menaquinol-8 --> 2 H + Menaquinone-8
2-Demethylmenaquinol-8 --> 2 H + 2-Demethylmenaquinone-8
H2O + ACCOA --> H + Acetate + COA
L-Glutamate + H2O --> 2-Oxoglutarate + Ammonium + 2 H
H[external] --> H

Additional energy dissipating reactions can be directly specified through config.energy.additional_energy_generating_reactions, which should be vector of COBREXA Reactions using the same metabolite name space as the model. Internally, the model is converted to a COBREXA StandardModel, so ensure that the appropriate accessors are defined for it. Since models use different name spaces, config.energy.energy_dissipating_metabolites is used to create the energy dissipating reactions. By default it uses the BiGG name space, but this will be changed to ChEBI in due course. If your model uses a different name space, then you have to change the values (NOT the keys) of config.energy.energy_dissipating_metabolites. Each energy dissipating reaction is added to the test only if all its associated metabolites are present. Any config.energy.optimizer_modifications to the solver are passed directly through to COBREXA's flux_balance_analysis function. All config.energy.boundary_reactions and config.energy.ignored_energy_reactions are deleted from an internal copy of model; this internal copy is used for analysis. Returns true if the model has no energy generating cycles.

Types and utilities