FROG

FBCModelTests.generate_frog_report — Method

generate_frog_report(
    model::COBREXA.SBMLModel,
    filename::String;
    report_dir,
    optimizer,
    workers,
    basefilename
)

FBCModelTests.generate_frog_report — Method

generate_frog_report(filename::String; kwargs...)

MEMOTE

FBCModelTests.is_model_charge_balanced — Method

is_model_charge_balanced(
    model::COBREXA.MetabolicModel;
    ignored_reactions
) -> Vector{String}

Iterates through all the reactions in model and checks if the charges across each reaction balance. Returns a list of reaction IDs that are charge unbalanced, which is empty if the test passes.

Optionally, pass a list of reactions to ignore in this process through ignored_reactions. It makes sense to include the biomass and exchange reactions in this list (default).

FBCModelTests.is_model_mass_balanced — Method

is_model_mass_balanced(
    model::COBREXA.MetabolicModel;
    ignored_reactions
) -> Vector{String}

Iterates through all the reactions in model and checks if the mass across each reaction balances. Returns a list of reaction IDs that are mass unbalanced, which is empty if the test passes.

Optionally, pass a list of reactions to ignore in this process through ignored_reactions. It makes sense to include the biomass and exchange reactions in this list (default).