FROG

FBCModelTests.generate_frog_report — Method

generate_frog_report(
    model::COBREXA.SBMLModel,
    filename::String;
    report_dir,
    optimizer,
    workers,
    basefilename
)

FBCModelTests.generate_frog_report — Method

generate_frog_report(filename::String; kwargs...)

MEMOTE

FBCModelTests.has_erroneous_energy_generating_cycles — Method

has_erroneous_energy_generating_cycles(
    model,
    optimizer;
    energy_dissipating_metabolites,
    additional_energy_generating_reactions,
    ignore_reactions,
    modifications,
    boundary_reactions
) -> Bool

Attempts to detect if the model contains any erroneous energy generating cycles by closing all exchange reactions and using flux balance analysis to maximize the sum of fluxes through a set of energy dissipating reactions. The flux sum should be zero if the model is free of energy generating reactions.

This function is based on Fritzemeier, Claus Jonathan, et al. "Erroneous energy-generating cycles in published genome scale metabolic networks: Identification and removal." PLoS computational biology (2017).

The energy dissipating reactions are based on the source paper, and include:

ATP + H2O --> ADP + H + Phosphate
CTP + H2O --> CDP + H + Phosphate
GTP + H2O --> GDP + H + Phosphate
UTP + H2O --> UDP + H + Phosphate
ITP + H2O --> IDP + H + Phosphate
NADH --> H + NAD
NADPH --> H + NADP
FADH2 --> 2 H + FAD
FMNH2 --> 2 H + FMN
Ubiquinol-8 --> 2 H + Ubiquinone-8
Menaquinol-8 --> 2 H + Menaquinone-8
2-Demethylmenaquinol-8 --> 2 H + 2-Demethylmenaquinone-8
H2O + ACCOA --> H + Acetate + COA
L-Glutamate + H2O --> 2-Oxoglutarate + Ammonium + 2 H
H[external] --> H

Additional energy dissipating reactions can be directly specified through additional_energy_generating_reactions, which should be vector of COBREXA Reactions using the same metabolite name space as the model. Internally, the model is converted to a COBREXA StandardModel, so ensure that the appropriate accessors are defined for it.

Since models use different name spaces, energy_dissipating_metabolites is used to create the energy dissipating reactions. By default it uses the BiGG name space, but this will be changed to ChEBI in due course. If your model uses a different name space, then you have to change the values (NOT the keys) of energy_dissipating_metabolites. Each energy dissipating reaction is added to the test only if all its associated metabolites are present. Any modifications to the solver are passed directly through to COBREXA's flux_balance_analysis function. All boundary_reactions and ignore_reactions are deleted from an internal copy of model; this internal copy is used for analysis.

Returns true if the model has energy generating cycles.

FBCModelTests.is_model_charge_balanced — Method

is_model_charge_balanced(
    model::COBREXA.MetabolicModel;
    ignored_reactions
) -> Vector{String}

Iterates through all the reactions in model and checks if the charges across each reaction balance. Returns a list of reaction IDs that are charge unbalanced, which is empty if the test passes.

Optionally, pass a list of reactions to ignore in this process through ignored_reactions. It makes sense to include the biomass and exchange reactions in this list (default).

FBCModelTests.is_model_mass_balanced — Method

is_model_mass_balanced(
    model::COBREXA.MetabolicModel;
    ignored_reactions
) -> Vector{String}

Iterates through all the reactions in model and checks if the mass across each reaction balances. Returns a list of reaction IDs that are mass unbalanced, which is empty if the test passes.

Optionally, pass a list of reactions to ignore in this process through ignored_reactions. It makes sense to include the biomass and exchange reactions in this list (default).