FROG

FBCModelTests.ResetObjective — Type

struct ResetObjective <: COBREXA.ModelWrapper

Fields

model::COBREXA.MetabolicModel

objective::SparseArrays.SparseVector{Float64, Int64}

FBCModelTests.frog_compare_reports — Method

frog_compare_reports(
    report_dir_a::String,
    report_dir_b::String
) -> Union{Test.FallbackTestSet, Test.DefaultTestSet}

A simple wrapper for comparing 2 previously generated reports in their respective directories.

FBCModelTests.frog_generate_report — Method

frog_generate_report(
    filename::String;
    report_dir,
    optimizer,
    workers,
    basefilename
)

A complete function for one-shot generation of FROG reports. Use frog_model_report, frog_metadata and frog_write_to_directory for finer control of the process.

FBCModelTests.frog_metadata — Method

frog_metadata(filename::String; optimizer, basefilename)

FBCModelTests.frog_model_report — Method

frog_model_report(
    model::COBREXA.SBMLModel;
    optimizer,
    workers
)

Generate FROGReportData for a model.

FBCModelTests.frog_objective_report — Method

frog_objective_report(
    sbml_model::COBREXA.SBMLModel,
    objective::String;
    optimizer,
    workers
)

Generate a FROGObjectiveReport containing the reproducibility data for a single objective in the SBML model.

FBCModelTests.frog_read_from_directory — Method

frog_read_from_directory(
    report_dir::String
) -> NamedTuple{(:metadata, :report), _A} where _A<:Tuple{Dict{String, String}, Dict}

Reverse of frog_write_to_directory.

FBCModelTests.frog_test_metadata_compatibility — Method

frog_test_metadata_compatibility(
    a::Dict{String, String},
    b::Dict{String, String}
)

FBCModelTests.frog_test_report_equality — Method

frog_test_report_equality(
    a::Dict{String, FROGObjectiveReport},
    b::Dict{String, FROGObjectiveReport};
    absolute_tolerance,
    relative_tolerance
) -> Union{Test.FallbackTestSet, Test.DefaultTestSet}

FBCModelTests.frog_write_to_directory — Method

frog_write_to_directory(
    r::Dict{String, FROGObjectiveReport},
    metadata::Dict{String, String};
    report_dir,
    basefilename
)

Write the contents of FROGReportData to the 4 TSV files as specified by FROG standard, and additionally write the metadata into the JSON file.

MEMOTE

FBCModelTests.model_has_no_erroneous_energy_generating_cycles — Method

model_has_no_erroneous_energy_generating_cycles(
    model,
    optimizer;
    config
) -> Bool

Attempts to detect if the model contains any erroneous energy generating cycles by closing all exchange reactions and using flux balance analysis to maximize the sum of fluxes through a set of energy dissipating reactions. The flux sum should be zero if the model is free of energy generating reactions.

This function is based on Fritzemeier, Claus Jonathan, et al. "Erroneous energy-generating cycles in published genome scale metabolic networks: Identification and removal." PLoS computational biology (2017).

The energy dissipating reactions are based on the source paper, and include:

ATP + H2O --> ADP + H + Phosphate
CTP + H2O --> CDP + H + Phosphate
GTP + H2O --> GDP + H + Phosphate
UTP + H2O --> UDP + H + Phosphate
ITP + H2O --> IDP + H + Phosphate
NADH --> H + NAD
NADPH --> H + NADP
FADH2 --> 2 H + FAD
FMNH2 --> 2 H + FMN
Ubiquinol-8 --> 2 H + Ubiquinone-8
Menaquinol-8 --> 2 H + Menaquinone-8
2-Demethylmenaquinol-8 --> 2 H + 2-Demethylmenaquinone-8
H2O + ACCOA --> H + Acetate + COA
L-Glutamate + H2O --> 2-Oxoglutarate + Ammonium + 2 H
H[external] --> H

Additional energy dissipating reactions can be directly specified through config.consistency.additional_energy_generating_reactions, which should be vector of COBREXA Reactions using the same metabolite name space as the model. Internally, the model is converted to a COBREXA StandardModel, so ensure that the appropriate accessors are defined for it.

Since models use different name spaces, config.consistency.energy_dissipating_metabolites is used to create the energy dissipating reactions. By default it uses the BiGG name space, but this will be changed to ChEBI in due course. If your model uses a different name space, then you have to change the values (NOT the keys) of config.consistency.energy_dissipating_metabolites. Each energy dissipating reaction is added to the test only if all its associated metabolites are present. Any config.consistency.optimizer_modifications to the solver are passed directly through to COBREXA's flux_balance_analysis function. All config.consistency.boundary_reactions and config.consistency.ignored_energy_reactions are deleted from an internal copy of model; this internal copy is used for analysis.

Returns true if the model has no energy generating cycles.

FBCModelTests.model_is_consistent — Method

model_is_consistent(model, optimizer; config) -> Bool

Determines if the model is stoichiometrically consistent. Note, stoichiometric consistency does not guarantee that mass balances must hold in the model. A more robust check is reactions_mass_unbalanced, but this works if not all metabolites have mass assigned to them.

Based on Gevorgyan, Albert, Mark G. Poolman, and David A. Fell. "Detection of stoichiometric inconsistencies in biomolecular models." Bioinformatics (2008).

Optionally ignore some reactions in this analysis by adding reaction IDs to config.consistency.consistency_ignored_reactions.

FBCModelTests.reactions_charge_unbalanced — Method

reactions_charge_unbalanced(model; config) -> Vector{String}

Iterates through all the reactions in model and checks if the charges across each reaction balance. Returns a list of reaction IDs that are charge unbalanced, which is empty if the test passes.

Optionally, use config.consistency.mass_ignored_reactions to pass a vector of reaction ids to ignore in this process. Internally biomass and exchang reactions are ignored.

FBCModelTests.reactions_mass_unbalanced — Method

reactions_mass_unbalanced(model; config) -> Vector{String}

Iterates through all the reactions in model and checks if the mass across each reaction balances. Returns a list of reaction IDs that are mass unbalanced, which is empty if the test passes.

Optionally, use config.consistency.charge_ignored_reactions to pass a vector of reaction ids to ignore in this process. Internally biomass and exchang reactions are ignored.

FBCModelTests.test_consistency — Method

test_consistency(
    model,
    optimizer;
    config
) -> Union{Test.FallbackTestSet, Test.DefaultTestSet}

Test if model is consistent by checking that:

the model is stoichiometrically consistent, tested with model_is_consistent

there are no energy generating cycles, tested with model_has_no_erroneous_energy_generating_cycles

the model is both mass and charge balanced, tested with reactions_charge_unbalanced and [reactions_mass_unbalanced]

Each function called in this test function can be called individually. The kwargs are forwarded as indicated by the prefix.

FBCModelTests.metabolites_are_duplicated — Method

metabolites_are_duplicated(model, m1, m2; config) -> Any

Test if metabolites m1 and m2 are different by comparing their config.metabolite.test_annotation field in the annotations of each metabolite. Note, if no annotations are present for one or both of the metabolites, then return true.

FBCModelTests.metabolites_duplicated_in_compartment — Method

metabolites_duplicated_in_compartment(
    model;
    config
) -> Dict{String, Set{String}}

Return a dictionary of metabolites that are duplicated in their compartment.

FBCModelTests.metabolites_medium_components — Method

metabolites_medium_components(
    model;
    config
) -> Vector{String}

Return a list of all boundary reactions that allow flux into the model and create a metabolite. Assume that boundary reactions only create a single metabolite. Use the testing config, default.metabolite.only_imported = false, to also return metabolites that can be produced by the model under default conditions.

FBCModelTests.metabolites_no_charge — Method

metabolites_no_charge(model; config) -> Any

List all metabolites without a charge. Use config.metabolite.charge_corner_cases to specify an extra case to check for charge's that are not properly assigned.

FBCModelTests.metabolites_no_formula — Method

metabolites_no_formula(model; config) -> Any

List all metabolites without a formula. Use config.metabolite.formula_corner_cases to specify an extra case to check for formula's that are not properly assigned.

FBCModelTests.metabolites_unique — Method

metabolites_unique(model; config) -> Set{String}

Return a list of unique metabolites in model. Uses metabolites_are_duplicated internally and forwards test_annotation to it. The latter argument is used to determine if two metabolites are the same by checking for any correspondence.

FBCModelTests.test_metabolites — Method

test_metabolites(
    model;
    config
) -> Union{Test.FallbackTestSet, Test.DefaultTestSet}

Test if the metabolites contained in the model:

are not duplicated, tested with metabolites_duplicated_in_compartment

all have a formula and charge associated with them, tested with metabolites_no_formula and metabolites_no_charge

the default medium of the cell is not empty, tested with metabolites_medium_components.

Each of the basic functions can be run independently. THe kwargs are forwarded as indicated by the prefix.

FBCModelTests.reactions_without_gpr — Method

reactions_without_gpr(model) -> Any

Return a list of reaction ids that do not have gene reaction rules (aka gene protein reaction associations).

FBCModelTests.atp_present_in_biomass — Method

atp_present_in_biomass(model; config) -> Any

Check if the biomass reaction consumes ATP and H₂O, and produces ADP, HO₄P, and H⁺. Each of these metabolites have a lookup table mapping them to the name space of the model, defined in config.biomass.growth_metabolites. These need to be set if you use anything other than the BiGG namespace.

FBCModelTests.biomass_missing_essential_precursors — Method

biomass_missing_essential_precursors(
    model;
    config
) -> Dict{String, Vector{String}}

Tests if each biomass reaction contains a set of essential precursors, listed in config.biomass.essential_precursors. Note, this function only works on a lumped biomass function.

FBCModelTests.find_blocked_biomass_precursors — Method

find_blocked_biomass_precursors(
    model,
    optimizer;
    config
) -> Dict{String, Vector{String}}

Check if the model can synthesize all of the biomass precursors in all of the biomass functions, except those listed in config.biomass.ignored_precursors in the default medium. Set any optimizer modifications with config.biomass.optimizer_modifications.

FBCModelTests.model_biomass_is_consistent — Method

model_biomass_is_consistent(model; config) -> Any

Check that the molar mass of each biomass reactions falls within [1 - 1e-3, 1 + 1e-6] by calling model_biomass_molar_mass internally.

FBCModelTests.model_biomass_molar_mass — Method

model_biomass_molar_mass(model; config) -> Any

For each biomass reaction, identified by model_biomass_reactions, calculate the molar weight of the reaction by summing the products of the associated metabolite coefficients with their molar masses.

FBCModelTests.model_biomass_reactions — Method

model_biomass_reactions(model; config) -> Set

Identify all the biomass reactions in the model using both sbo annotations, as well biomass strings typically contained in their reaction IDs. Use config.biomass.biomass_strings to update the list of strings to look for.

FBCModelTests.model_has_atpm_reaction — Method

model_has_atpm_reaction(
    model;
    config
) -> Union{Missing, Bool}

Check if model has an ATP maintenance reaction built in (also called a non-growth associated maintenance cost). Looks for reaction annotations corresponding to the sbo maintenance term, or looks for reaction ids that contain the strings listed in config.biomass.atpm_strings.

FBCModelTests.model_solves_in_default_medium — Method

model_solves_in_default_medium(
    model,
    optimizer;
    config
) -> Any

Check if the model can be solved under default conditions and yield a reasonable growth rate. Here reasonable is set via config.biomass.minimum_growth_rate and config.biomass.maximum_growth_rate. Optionally, pass optimization modifications to the solver through config.biomass.optimizer_modifications.

FBCModelTests._find_orphan_or_deadend_metabolites — Method

_find_orphan_or_deadend_metabolites(
    model;
    consumed
) -> Vector{String}

Helper function to find orphan or deadend metabolites. Specify consumed=true to consider orphan metabolites or false to consider deadend metabolites. Set complete_medium=true to open all boundary reactions to simulate a complete medium.

FBCModelTests.find_all_universally_blocked_reactions — Method

find_all_universally_blocked_reactions(
    model,
    optimizer;
    config
) -> Any

Make all boundary reactions reversible and run FVA on the model to find all reactions that are universally blocked. Optimizer modifications can be passed through config.network.optimizer_modifications

FBCModelTests.find_complete_medium_orphans_and_deadends — Method

find_complete_medium_orphans_and_deadends(
    model,
    optimizer;
    config
) -> Dict{Symbol, Vector{String}}

Return a dictionary mapping orphan and deadend metabolites that occur in the model in complete medium. Complete medium is modeled by opening all the bounday reactions. For each metabolite FBA is run with a temporary reaction either consuming or producing the metabolite in question. At minimum a flux of config.network.minimum_metabolite_flux must be attained for the metabolite to pass the test.

FBCModelTests.find_cycle_reactions — Method

find_cycle_reactions(
    model,
    optimizer;
    config
) -> Set{String}

Find all reactions that participate in stoichiometrically balanced cycles by closing all boundary reactions and running fva on the resultant model.

FBCModelTests.find_deadend_metabolites — Method

find_deadend_metabolites(model) -> Vector{String}

Find all metabolites that can only (excludes reversible reactions) be produced in the model by inspecting the stoichiometric matrix.

FBCModelTests.find_orphan_metabolites — Method

find_orphan_metabolites(model) -> Vector{String}

Find all metabolites that can only (excludes reversible reactions) be consumed in the model by inspecting the stoichiometric matrix.

FBCModelTests.stoichiometric_matrix_is_well_conditioned — Method

stoichiometric_matrix_is_well_conditioned(
    model;
    config
) -> Any

Test if the stoichiometric matrix is well conditioned by determining if stoichiometric_max_min_ratio is less than 10⁹ (which can be set in config.network.condition_number).

FBCModelTests.stoichiometric_max_min_ratio — Method

stoichiometric_max_min_ratio(model) -> Any

Return the ratio of the absolute maximum and minimum value of the nonzero coefficients in the stoichiometric matrix of model.

FBCModelTests.all_unannotated_genes — Method

all_unannotated_genes(model::COBREXA.MetabolicModel) -> Any

Checks if every gene has an annotation and returns an array of genes which do not have annotatons.

Types and utilities

FBCModelTests.ObjectiveValue — Type

FBCModelTests.BiomassConfig — Type

mutable struct BiomassConfig

Parameters used by the biomass tests.

Fields

biomass_strings::Vector{String}

atpm_strings::Vector{String}

growth_metabolites::Dict{String, String}

minimum_growth_rate::Float64

maximum_growth_rate::Float64

ignored_precursors::Vector{String}

essential_precursors::Dict{String, String}

optimizer_modifications::Vector{Function}

FBCModelTests.ConsistencyConfig — Type

mutable struct ConsistencyConfig

Parameters used by the consistency tests.

Fields

mass_ignored_reactions::Vector{String}

charge_ignored_reactions::Vector{String}

consistency_ignored_reactions::Vector{String}

energy_dissipating_metabolites::Dict{String, String}

additional_energy_generating_reactions::Vector{COBREXA.Reaction}

ignored_energy_reactions::Vector{String}

optimizer_modifications::Vector{Function}

FBCModelTests.FROGMetadata — Type

mutable struct Dict{String, String} <: AbstractDict{String, String}

FBCModelTests.FROGObjectiveReport — Type

struct FROGObjectiveReport

Fields

optimum::Union{Nothing, Float64}

reactions::Dict{String, FROGReactionReport}

gene_deletions::Dict{String, Union{Nothing, Float64}}

FBCModelTests.FROGReactionReport — Type

struct FROGReactionReport

Fields

flux::Union{Nothing, Float64}

variability_min::Union{Nothing, Float64}

variability_max::Union{Nothing, Float64}

deletion::Union{Nothing, Float64}

FBCModelTests.FROGReportData — Type

mutable struct Dict{String, FROGObjectiveReport} <: AbstractDict{String, FROGObjectiveReport}

FBCModelTests.MemoteConfig — Type

mutable struct MemoteConfig

A grouping of parameters used by the metabolic testing infrastructure.

Fields

metabolite::MetaboliteConfig

consistency::ConsistencyConfig

biomass::BiomassConfig

network::NetworkConfig

FBCModelTests.MetaboliteConfig — Type

mutable struct MetaboliteConfig

Parameters used by the metabolite tests.

Fields

formula_corner_cases::Vector{String}

charge_corner_cases::Vector{Int64}

medium_only_imported::Bool

test_annotation::String

FBCModelTests._has_sensible_gpr — Method

_has_sensible_gpr(model, rid) -> Any

Internal helper function that determines if a reaction has a gene reaction rule and that each gene in the rule is contained in the model.

FBCModelTests._probably_transport_reaction — Method

_probably_transport_reaction(
    model,
    rid,
    test_annotation
) -> Bool

Determine if a reaction is probably a transport reaction by checking if:

it has sbo annotations corresponding to a transport reaction

the reaction contains metabolites from at least 2 different compartments

if at least 1 metabolite does not undergo a chemical transformation (via formula or annotation checks)

Note, PTS type transport reactions will be missed if they do not have sbo annotations. This test may yield false negatives.

FBCModelTests.to_element — Method

to_element(x::String) -> PeriodicTable.Element

Return the chemical element of x.