FROG

FBCModelTests.ResetObjective — Type

struct ResetObjective <: COBREXA.ModelWrapper

Fields

model::COBREXA.MetabolicModel

objective::SparseArrays.SparseVector{Float64, Int64}

FBCModelTests.frog_compare_reports — Method

frog_compare_reports(
    report_dir_a::String,
    report_dir_b::String
) -> Union{Test.FallbackTestSet, Test.DefaultTestSet}

A simple wrapper for comparing 2 previously generated reports in their respective directories.

FBCModelTests.frog_generate_report — Method

frog_generate_report(
    filename::String;
    report_dir,
    optimizer,
    workers,
    basefilename
)

A complete function for one-shot generation of FROG reports. Use frog_model_report, frog_metadata and frog_write_to_directory for finer control of the process.

FBCModelTests.frog_metadata — Method

frog_metadata(filename::String; optimizer, basefilename)

FBCModelTests.frog_model_report — Method

frog_model_report(
    model::COBREXA.SBMLModel;
    optimizer,
    workers
)

Generate FROGReportData for a model.

FBCModelTests.frog_objective_report — Method

frog_objective_report(
    sbml_model::COBREXA.SBMLModel,
    objective::String;
    optimizer,
    workers
)

Generate a FROGObjectiveReport containing the reproducibility data for a single objective in the SBML model.

FBCModelTests.frog_read_from_directory — Method

frog_read_from_directory(
    report_dir::String
) -> NamedTuple{(:metadata, :report), _A} where _A<:Tuple{Dict{String, String}, Dict}

Reverse of frog_write_to_directory.

FBCModelTests.frog_test_metadata_compatibility — Method

frog_test_metadata_compatibility(
    a::Dict{String, String},
    b::Dict{String, String}
)

FBCModelTests.frog_test_report_equality — Method

frog_test_report_equality(
    a::Dict{String, FBCModelTests.FROGObjectiveReport},
    b::Dict{String, FBCModelTests.FROGObjectiveReport};
    absolute_tolerance,
    relative_tolerance
) -> Union{Test.FallbackTestSet, Test.DefaultTestSet}

FBCModelTests.frog_write_to_directory — Method

frog_write_to_directory(
    r::Dict{String, FBCModelTests.FROGObjectiveReport},
    metadata::Dict{String, String};
    report_dir,
    basefilename
)

Write the contents of FROGReportData to the 4 TSV files as specified by FROG standard, and additionally write the metadata into the JSON file.

MEMOTE

FBCModelTests.has_erroneous_energy_generating_cycles — Method

has_erroneous_energy_generating_cycles(
    model,
    optimizer;
    energy_dissipating_metabolites,
    additional_energy_generating_reactions,
    ignore_reactions,
    modifications,
    boundary_reactions
) -> Bool

Attempts to detect if the model contains any erroneous energy generating cycles by closing all exchange reactions and using flux balance analysis to maximize the sum of fluxes through a set of energy dissipating reactions. The flux sum should be zero if the model is free of energy generating reactions.

This function is based on Fritzemeier, Claus Jonathan, et al. "Erroneous energy-generating cycles in published genome scale metabolic networks: Identification and removal." PLoS computational biology (2017).

The energy dissipating reactions are based on the source paper, and include:

ATP + H2O --> ADP + H + Phosphate
CTP + H2O --> CDP + H + Phosphate
GTP + H2O --> GDP + H + Phosphate
UTP + H2O --> UDP + H + Phosphate
ITP + H2O --> IDP + H + Phosphate
NADH --> H + NAD
NADPH --> H + NADP
FADH2 --> 2 H + FAD
FMNH2 --> 2 H + FMN
Ubiquinol-8 --> 2 H + Ubiquinone-8
Menaquinol-8 --> 2 H + Menaquinone-8
2-Demethylmenaquinol-8 --> 2 H + 2-Demethylmenaquinone-8
H2O + ACCOA --> H + Acetate + COA
L-Glutamate + H2O --> 2-Oxoglutarate + Ammonium + 2 H
H[external] --> H

Additional energy dissipating reactions can be directly specified through additional_energy_generating_reactions, which should be vector of COBREXA Reactions using the same metabolite name space as the model. Internally, the model is converted to a COBREXA StandardModel, so ensure that the appropriate accessors are defined for it.

Since models use different name spaces, energy_dissipating_metabolites is used to create the energy dissipating reactions. By default it uses the BiGG name space, but this will be changed to ChEBI in due course. If your model uses a different name space, then you have to change the values (NOT the keys) of energy_dissipating_metabolites. Each energy dissipating reaction is added to the test only if all its associated metabolites are present. Any modifications to the solver are passed directly through to COBREXA's flux_balance_analysis function. All boundary_reactions and ignore_reactions are deleted from an internal copy of model; this internal copy is used for analysis.

Returns true if the model has energy generating cycles.

FBCModelTests.is_consistent — Method

is_consistent(
    model,
    optimizer;
    boundary_reactions,
    ignored_reactions
) -> Bool

Determines if the model is stiochiometrically consistent. Note, stoichiometric consistency does not guarantee that mass balances must hold in the model. A more robust check is is_model_mass_balanced, but this works if not all metabolites have mass assigned to them.

Based on Gevorgyan, Albert, Mark G. Poolman, and David A. Fell. "Detection of stoichiometric inconsistencies in biomolecular models." Bioinformatics (2008).

FBCModelTests.is_model_charge_balanced — Method

is_model_charge_balanced(
    model::COBREXA.MetabolicModel;
    ignored_reactions
) -> Vector{String}

Iterates through all the reactions in model and checks if the charges across each reaction balance. Returns a list of reaction IDs that are charge unbalanced, which is empty if the test passes.

Optionally, pass a list of reactions to ignore in this process through ignored_reactions. It makes sense to include the biomass and exchange reactions in this list (default).

FBCModelTests.is_model_mass_balanced — Method

is_model_mass_balanced(
    model::COBREXA.MetabolicModel;
    ignored_reactions
) -> Vector{String}

Iterates through all the reactions in model and checks if the mass across each reaction balances. Returns a list of reaction IDs that are mass unbalanced, which is empty if the test passes.

Optionally, pass a list of reactions to ignore in this process through ignored_reactions. It makes sense to include the biomass and exchange reactions in this list (default).

Types and utilities

FBCModelTests.ObjectiveValue — Type

FBCModelTests.FROGMetadata — Type

mutable struct Dict{String, String} <: AbstractDict{String, String}

FBCModelTests.FROGObjectiveReport — Type

struct FROGObjectiveReport

Fields

optimum::Union{Nothing, Float64}

reactions::Dict{String, FBCModelTests.FROGReactionReport}

gene_deletions::Dict{String, Union{Nothing, Float64}}

FBCModelTests.FROGReactionReport — Type

struct FROGReactionReport

Fields

flux::Union{Nothing, Float64}

variability_min::Union{Nothing, Float64}

variability_max::Union{Nothing, Float64}

deletion::Union{Nothing, Float64}

FBCModelTests.FROGReportData — Type

mutable struct Dict{String, FBCModelTests.FROGObjectiveReport} <: AbstractDict{String, FBCModelTests.FROGObjectiveReport}